requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
blinker==1.4
boto3==1.20.24
botocore==1.23.24
bravado==11.0.3
bravado-core==5.17.0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
cycler==0.11.0
dataclasses==0.8
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dgl==0.7.2
entrypoints==0.3
fonttools==4.28.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
googledrivedownloader==0.4
greenlet==1.1.2
idna==3.1
importlib-metadata==4.8.2
ipykernel==6.4.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonref==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.2
littleutils==0.2.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monotonic==1.5
msgpack==1.0.3
munkres==1.1.4
nbclient==0.5.3
nbconvert==6.3.0
nbformat==5.1.3
neptune-client==0.14.0
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.6
numpy==1.21.2
oauthlib==3.1.1
ogb==1.3.2
olefile==0.46
outdated==0.0.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
prometheus-client==0.12.0
prompt-toolkit==3.0.20
psutil==5.8.0
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-louvain==0.15
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.10.0
reportlab==3.5.68
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.1
Send2Trash==1.8.0
setuptools==58.0.4
simplejson==3.17.6
sip==4.19.13
six==1.16.0
smmap==3.0.5
SQLAlchemy==1.4.31
swagger-spec-validator==2.7.4
terminado==0.9.4
testpath==0.5.0
threadpoolctl==3.0.0
torch==1.10.0
torch-cluster==1.5.9
torch-geometric==2.0.2
torch-scatter==2.0.9
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torchaudio==0.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
wheel==0.37.0
widgetsnbextension==3.5.1
yacs==0.1.6
zipp==3.6.0
